+++ Smart Calculator/task/src/calculator/operations/Operation.java
@@ -1,6 +1,8 @@
 package calculator.operations;
 
+import java.math.BigInteger;
+
 public abstract class Operation {
     public Priority priority;
-    public abstract int count(int params, int b);
+    public abstract BigInteger count(BigInteger params, BigInteger b);
 }
